backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
websockets==12.0
httpx==0.26.0
solders==0.21.0
resend==0.8.0
pytest==7.4.4
pytest-asyncio==0.23.3
pynacl==1.5.0
base58==2.1.1
